# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JagoEvent Ticketing API",
        "description": "REST API untuk aplikasi JagoEvent — marketplace tiket event. Digunakan oleh aplikasi mobile Flutter untuk browse events, order tickets, manage vendor, dan validasi tiket.",
        "contact": {
            "name": "JagoEvent Support",
            "email": "support@JagoEvent.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "JagoEvent API"
        }
    ],
    "paths": {
...
        "/order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Buat order tiket baru",
                "description": "Membuat order baru, book tiket, dan generate payment URL via Midtrans.",
                "operationId": "8cf9152f902d1d264fa3f5d31ab1c685",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id",
                                    "order_details",
                                    "quantity",
                                    "event_date"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "sku_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "event_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "total_price": {
                                                    "type": "integer"
                                                },
                                                "event_date": {
                                                    "type": "string"
                                                },
                                                "status_payment": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "payment_url": {
                                                    "description": "Midtrans Snap redirect URL",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "user": {
                                                    "type": "object"
                                                },
                                                "orderItems": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error (tiket habis, validation error, dll)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update status payment order",
                "operationId": "1ef921eb3918724d06805ac814d382ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_payment"
                                ],
                                "properties": {
                                    "status_payment": {
                                        "type": "string",
                                        "example": "success",
                                        "enum": [
                                            "pending",
                                            "success",
                                            "cancel"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/orders/user/{id}/vendor": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders untuk vendor (semua event milik vendor)",
                "operationId": "0c6d709dfd1544bcc1fe60c544628fb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID (vendor)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar order untuk event vendor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Sanctum token (without Bearer prefix)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}